alta pedido v3 (creacion de DAOs y Managers)

+++ src/main/java/objects/Persona.java
@@ -43,7 +43,7 @@ public abstract class Persona {
         return telefono;
     }
 
-    public void setTelefono(int telefono) {
+    public void setTelefono(long telefono) {
         this.telefono = telefono;
     }
